Move test data

+++ src/back_end/pose_graph_manager_node.cpp
@@ -20,7 +20,7 @@ int main(int argc, char **argv) {
   node->declare_parameter<int>("pose_graph_manager_process_period_ms", 1000);
   node->declare_parameter<int>("pose_graph_optimization_loop_period_ms", 100);
   node->declare_parameter<int>("max_waiting_time_sec", 100);
-  node->declare_parameter<int>("heartbeat_period_sec", 1.0);
+  node->declare_parameter<double>("heartbeat_period_sec", 1.0);
 
   PoseGraphManager manager(node);
